unnecessary_clone: split rustfixable lint out into separate test
